requirements.txt
flask
gunicorn
python-dotenv
google-generativeai
pymupdf
reportlab
werkzeug
